main_project/requirements.txt
[[source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"

[packages]
flask = "*"
flask-sqlalchemy = "*"
flask-marshmallow = "*"
marshmallow-sqlalchemy = "*"
pymysql = "*"

[dev-packages]
beautifulsoup4~=4.10.0
pandas~=1.3.5

[requires]
python_version = "3.9"
Flask~=2.0.2
SQLAlchemy~=1.4.29
requests~=2.27.1